add tauri setup file
